TST: minimize_ipopt: combined tests from SciPy, added license
